Allow safe approximations of _notid.
The safe approximation includes some strings which are actually identities,
but guarantees that all nonidentities are included.

This approximation would be better if we visited states in order of
cumulative frequency, eg using weights on the edges.  A breadth-first
sort of the node would roughly trim the approximation by input string
length.